+++ plutus-laws/src/Test/Tasty/Plutus/Laws.hs
@@ -41,6 +41,10 @@ module Test.Tasty.Plutus.Laws (
   plutusEqLawsWith,
   plutusEqLawsDirect,
   plutusEqLawsDirectWith,
+  plutusEqLawsSubstitution,
+  plutusEqLawsSubstitutionWith,
+  plutusEqLawsSubstitutionDirect,
+  plutusEqLawsSubstitutionDirectWith,
 
   -- *** Ord
   plutusOrdLaws,
@@ -81,12 +85,18 @@ import Test.QuickCheck (
 import Test.QuickCheck.Arbitrary (
   Arbitrary (arbitrary, shrink),
   Arbitrary1 (liftArbitrary, liftShrink),
+  CoArbitrary,
+  liftArbitrary2,
+  liftShrink2,
  )
+import Test.QuickCheck.Function (Fun, Function, applyFun)
 import Test.QuickCheck.Gen (Gen)
+import Test.QuickCheck.Poly (A)
 import Test.Tasty (TestTree)
 import Test.Tasty.Plutus.Arbitrary (
   Entangled (Disentangled, Entangled),
   Entangled3 (Disentangled3, Entangled3),
+  PA,
   Pair (Pair),
   Triple (Triple),
  )
@@ -189,7 +199,7 @@ dataLawsWith gen shr =
  This uses a technique to avoid coverage issues arising from a low likelihood
  of independently generating identical values. This mostly affects those types
  which have a large, or infinite, number of inhabitants: if your type is
- finite and small, this will actually have the _opposite_ effect, as it would
+ finite and small, this will actually have the /opposite/ effect, as it would
  skew the balance of comparisons the /other/ way.
 
  To assist with this, coverage checking is built in: if you are seeing errors
@@ -333,6 +343,142 @@ plutusEqLawsDirectWith gen shr =
         "precondition satisfied"
         $ ((x PlutusTx.== y) && (y PlutusTx.== z)) === (x PlutusTx.== z)
 
+{- | Checks that the substitution property holds for @a@'s 'PlutusTx.Eq'
+ instance; specifically, that for any pure function @f@, if @x == y@, then @f
+ x == f y@.
+
+ = Note
+
+ This uses a technique to avoid coverage issues arising from a low likelihood
+ of independently generating identical values. This mostly affects those types
+ which have a large, or infinite, number of inhabitants: if your type is
+ finite and small, this will actually have the /opposite/ effect, as it would
+ skew the balance of comparisons the /other/ way.
+
+ To assist with this, coverage checking is built in: if you are seeing errors
+ due to coverage, try using 'plutusEqLawsSubstitutionDirect' instead.
+
+ @since 2.1
+-}
+plutusEqLawsSubstitution ::
+  forall (a :: Type).
+  (Typeable a, Arbitrary a, Show a, PlutusTx.Eq a, Function a, CoArbitrary a) =>
+  TestTree
+plutusEqLawsSubstitution = plutusEqLawsSubstitutionWith @a arbitrary shrink
+
+{- | As 'plutusEqLawsSubstitution', but with explicit generator and shrinker.
+
+ = Note
+
+ As this function (like 'plutusEqLawsSubstitution') uses a technique to avoid
+ coverage issues, if your generator and shrinker are restricted (especially to
+ a small, finite subset of the whole type), you may get test failures due to
+ poor coverage. If this happens, either use
+ 'plutusEqLawsSubstitutionDirectWith', or change the generator and shrinker to
+ be less constrained.
+
+ @since 2.1
+-}
+plutusEqLawsSubstitutionWith ::
+  forall (a :: Type).
+  (Typeable a, Show a, PlutusTx.Eq a, Function a, CoArbitrary a) =>
+  Gen a ->
+  (a -> [a]) ->
+  TestTree
+plutusEqLawsSubstitutionWith gen shr =
+  testProperties
+    ("Plutus Eq substitutability for " <> typeName @a)
+    [
+      ( "if x == y then f x == f y"
+      , forAllShrinkShow genWithFun shrinkWithFun ppShow propSub
+      )
+    ]
+  where
+    genWithFun :: Gen (Entangled a, Fun a PA)
+    genWithFun = liftArbitrary2 (liftArbitrary gen) arbitrary
+    shrinkWithFun ::
+      (Entangled a, Fun a PA) ->
+      [(Entangled a, Fun a PA)]
+    shrinkWithFun (ent, fn) =
+      liftShrink2 (liftShrink shr) shrink (ent, fn)
+    propSub :: (Entangled a, Fun a PA) -> Property
+    propSub (ent, fn) = checkCoverage
+      . cover 50.0 (knownEntangled ent) "precondition known satisfied"
+      $ case ent of
+        Entangled x y -> applyFun fn x === applyFun fn y
+        Disentangled x y ->
+          if x PlutusTx.== y
+            then (applyFun fn x PlutusTx.== applyFun fn y) === True
+            else property True
+
+{- | Checks that the substitution property holds for @a@'s 'PlutusTx.Eq'
+ instance; specifically, that for any pure function @f@, if @x == y@, then @f
+ x == f y@.
+
+ = Note
+
+ This function tests the substitution property directly, without relying on
+ the coverage-improving technique used in 'plutusEqLawsSubstitution'. This
+ works if @a@ as a type is both finite and has small cardinality, as in that
+ situation, the distribution of successful and unsuccessful preconditions will
+ roughly match. However, if @a@ is infinite or large, this will cause a skewed
+ case distribution, which will produce misleading results.
+
+ To assist with this, coverage checking is built in to the property this
+ checks; if you find that your coverage is too low, use
+ 'plutusEqLawsSubstitution' instead.
+
+ @since 2.1
+-}
+plutusEqLawsSubstitutionDirect ::
+  forall (a :: Type).
+  (Typeable a, Arbitrary a, Show a, PlutusTx.Eq a, Function a, CoArbitrary a) =>
+  TestTree
+plutusEqLawsSubstitutionDirect =
+  plutusEqLawsSubstitutionDirectWith @a arbitrary shrink
+
+{- | As 'plutusEqLawsSubstitutionDirect', but with explicit generator and
+ shrinker.
+
+ = Note
+
+ As this function (like 'plutusEqLawsSubstitutionDirect') tests the
+ substutition property directly, either @a@ must be both finite and small, or
+ the generator and shrinker passed to this function must produce only a small
+ and finite subset of the type. See the caveats on the use of
+ 'plutusEqLawsSubstitutionDirect' as to why.
+
+ To assist with this, coverage checking is built in: if you find that your
+ coverage is too low, use 'plutusEqLawsSubstitution' instead.
+
+ @since 2.1
+-}
+plutusEqLawsSubstitutionDirectWith ::
+  forall (a :: Type).
+  (Typeable a, Show a, PlutusTx.Eq a, Function a, CoArbitrary a) =>
+  Gen a ->
+  (a -> [a]) ->
+  TestTree
+plutusEqLawsSubstitutionDirectWith gen shr =
+  testProperties
+    ("Plutus Eq substitutability for " <> typeName @a)
+    [
+      ( "if x == y then f x == f y"
+      , forAllShrinkShow genWithFun shrinkWithFun ppShow propSub
+      )
+    ]
+  where
+    genWithFun :: Gen (Pair a, Fun a A)
+    genWithFun = liftArbitrary2 (liftArbitrary gen) arbitrary
+    shrinkWithFun :: (Pair a, Fun a A) -> [(Pair a, Fun a A)]
+    shrinkWithFun (p, fn) = liftShrink2 (liftShrink shr) shrink (p, fn)
+    propSub :: (Pair a, Fun a A) -> Property
+    propSub (Pair x y, fn) =
+      cover 50.0 (x PlutusTx.== y) "precondition satisfied" $
+        if x PlutusTx.== y
+          then applyFun fn x === applyFun fn y
+          else property True
+
 {- | Checks that the 'PlutusTx.Ord' instance for @a@ is a total order.
 
  = Note
@@ -341,7 +487,7 @@ plutusEqLawsDirectWith gen shr =
  of independently generating identical values (specifically in the tests for
  antisymmetry). This mostly affects those types which have a large, or
  infinite, number of inhabitants: if your type is finite and small, this will
- actually have the _opposite_ effect, as it would skew the comparisons the
+ actually have the /opposite/ effect, as it would skew the comparisons the
  /other/ way.
 
  To assist with this, coverage checking is built in: if you are seeing errors