+++ plutus-numeric/src/PlutusTx/NatRatio.hs
@@ -32,9 +32,22 @@ module PlutusTx.NatRatio (
   Internal.recip,
   Internal.toRational,
 
-  -- ** NatRatio
+  -- ** NatRatioSchema
   Internal.NatRatioSchema (NatRatioSchema),
+
+  -- ** Conversion
+  natRatioToRational,
 ) where
 
 import PlutusTx.NatRatio.Internal qualified as Internal
 import PlutusTx.NatRatio.QQ qualified as QQ
+import PlutusTx.Numeric.Extra (addExtend)
+import PlutusTx.Ratio qualified as Ratio
+
+{- | The same as 'addExtend', but specialized for the 'Internal.NatRatio' to
+ 'Ratio.Rational' case.
+
+ @since 2.1
+-}
+natRatioToRational :: Internal.NatRatio -> Ratio.Rational
+natRatioToRational = addExtend